+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.16.7-dev0
+
+### Enhancements
+- **Add image_alt_mode to partition_html** Adds an `image_alt_mode` parameter to `partition_html()` to control how alt text is extracted from images in HTML documents. The parameter can be set to `to_text` to extract alt text as text from <img> html tags
+
+### Features
+
+### Fixes
+
 ## 0.16.6
 
 ### Enhancements